requirements.txt
fastapi
uvicorn
tensorflow==2.9.1
tensorflow-hub==0.12.0
numpy<2
pillow
python-multipart
